+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
-## next
+## [0.2.0] - 2025-12-15
+
+[0.2.0]: https://github.com/langston-barrett/lun/releases/tag/v0.2.0
 
 - `tool.files` now requires a list of globs
 - Add `tool.ignore` for ignoring a list of globs